+++ source/glest_game/world/world.cpp
@@ -1164,7 +1164,7 @@ void World::addAttackEffects(const Unit *unit) {
 
 //returns the nearest unit that can store a type of resource given a position and a faction
 Unit *World::nearestStore(const Vec2i &pos, int factionIndex, const ResourceType *rt) {
-    double currDist = infinity;
+	float currDist = infinity;
     Unit *currUnit= NULL;
 
     if(factionIndex >= getFactionCount()) {
@@ -1174,7 +1174,7 @@ Unit *World::nearestStore(const Vec2i &pos, int factionIndex, const ResourceType
     for(int i=0; i < getFaction(factionIndex)->getUnitCount(); ++i) {
 		Unit *u= getFaction(factionIndex)->getUnit(i);
 		if(u != NULL) {
-			double tmpDist= u->getPos().dist(pos);
+			float tmpDist= u->getPos().dist(pos);
 			if(tmpDist < currDist &&  u->getType() != NULL && u->getType()->getStore(rt) > 0 && u->isOperative()) {
 				currDist= tmpDist;
 				currUnit= u;
@@ -2348,7 +2348,7 @@ void World::exploreCells(const Vec2i &newPos, int sightRange, int teamIndex) {
 				}
 
 				//explore
-				double posLength = currRelPos.length();
+				float posLength = currRelPos.length();
 				//if(Vec2i(0).dist(currRelPos) < surfSightRange + indirectSightRange + 1) {
 				if(posLength < surfSightRange + indirectSightRange + 1) {
                     sc->setExplored(teamIndex, true);
